actualización de data de hito 2 falló conexión data
