+++ public/covid19/assets/js/profe_dataUSA.js
@@ -1,3 +1,14 @@
+//capturando data de USA
+/* fetch('http://localhost:3000/api/total')
+  .then(res => res.json())
+  .then(data => dataUSA(data))
+  .catch(err => console.log(err, "Ocurrio un error"))
+ */
+/* const dataUSA = (data) => {
+  console.log("mostrando la data USA:", data[0]);
+  //getData();
+};
+ */
 //en proceso JWT aun pero por buen camino 
 const PostData = async (email, password) => {
   try {
@@ -5,27 +16,30 @@ const PostData = async (email, password) => {
     console.log('PASSWORD', password)
     let payload = { email, password }
     const response = await fetch('http://localhost:3000/api/login',
-      {
-        method: "POST",
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(payload)
-
-      });
-        // console.log('TOKEN FRONT', responde.json())
+    {
+      method: "POST" ,
+      headers: {'Content-Type': 'application/json'},
+      body: JSON.stringify(payload)
+      
+    });
+    console.log('TOKEN FRONT', await response.json())
     const token = await response.json()
     console.log('TOKEN FRONT', token)
-    // if (token) {
-      console.log("acceso exitoso")
+   // if (token) {
+      console.log("acceso exitoso");
       return token
-    // } else (console.log("acceso denegado"))
-  } catch (error) {
+
+   // }else(console.log("acceso denegado"))
+  }catch(error){
     console.log(error);
   }
+  
 };
 
 const getData = async (jwt) => {
   try {
-    let {token} = jwt;
+    let { token } = jwt;
+
     const response = await fetch('http://localhost:3000/api/country/usa',
       {
         method: 'GET',
@@ -38,7 +52,6 @@ const getData = async (jwt) => {
   } catch (err) {
     console.log(`Error: ${err}`)
   }
-  console.log(getData())
 };
 
 document.getElementById('formulario').addEventListener('submit', async (event) => {
@@ -47,7 +60,5 @@ document.getElementById('formulario').addEventListener('submit', async (event) =
   const password = document.getElementById('exampleInputPassword1').value
   const JWT = await PostData(email, password)
   const posts = await getData(JWT)
-  console.log(posts);
+  console.log(posts)
 })
-
-export default getData();
