--- public/covid19/assets/js/render.js
@@ -1,2 +1,0 @@
-import { dataModelOrdenadaCortada } from "./jsData.js";
-console.log("Data importada de jsDATA :",dataModelOrdenadaCortada);
