+++ public/covid19/usah1.html
@@ -77,16 +77,17 @@
   <form id="formulario">
     <div class="form-group">
     <label for="exampleInputEmail1" id="emailpasswordtext" >Ingrese su correo electronico</label>
-    <input type="email" class="form-control" id="exampleInputEmail1"aria-describedby="emailHelp" placeholder="Ingrese  aqui correo electronico">
+    <input type="email" class="form-control" id="exampleInputEmail1"aria-describedby="emailHelp" placeholder="Ingrese  aqui correo electronico" value="Sincere@april.biz">
     </div>
     <div class="form-group">
     <label for="exampleInputPassword1" id="emailpasswordtext">Ingrese su contraseña</label>
-    <input type="password" class="form-control" id="exampleInputPassword1" placeholder="Ingrese aqui su contraseña">
+    <input type="password" class="form-control" id="exampleInputPassword1" placeholder="Ingrese aqui su contraseña" value="123">
     </div>
     <button type="submit" class="btn btn-primary ">Iniciar Sesion</button>
     </form>
   </div>
     
+  <div class="col-12" id="GraficoUSA" style="height:800px"></div>
 
   <!-- jQuery, Popper.js, Bootstrap JS -->
   <script src="https://code.jquery.com/jquery-3.6.0.js" integrity="sha256-H+K7U5CnXl1h5ywQfKtSj8PCmoN9aaq30gDh27Xc0jk="
@@ -102,5 +103,6 @@
   <!-- Se lee despues de haber leido jQuery, Popper.js, Bootstrap JS. Viene el script del proyecto -->
 
   <script src="./assets/js/dataUSA.js"></script>
+  <script type="module" src="./assets/js/jsChartUsa.js"></script>
 </body>
 </html>
